requirements.txt
Pillow
pycairo
pyGObject
